Refactor ToolUse into separate module
