+++ README.md
@@ -90,7 +90,8 @@ import json
 from dataclasses import dataclass
 
 from ai_agent_toolbox import Toolbox
-from ai_agent_toolbox.parser_event import ParserEvent, ToolUse
+from ai_agent_toolbox.parser_event import ParserEvent
+from ai_agent_toolbox.tool_use import ToolUse
 
 
 @dataclass
