+++ ai_agent_toolbox/parsers/xml/tool_parser.py
@@ -1,6 +1,7 @@
 import uuid
 from typing import Optional, List, Dict
-from ai_agent_toolbox.parser_event import ParserEvent, ToolUse
+from ai_agent_toolbox.parser_event import ParserEvent
+from ai_agent_toolbox.tool_use import ToolUse
 
 class ToolParserState:
     WAITING_FOR_NAME = "waiting_for_name"
