+++ ai_agent_toolbox/tool_use.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass, field
+from typing import Any, Dict
+
+@dataclass
+class ToolUse:
+    name: str
+    args: Dict[str, Any] = field(default_factory=dict)
