+++ tests/test_toolbox.py
@@ -4,7 +4,8 @@ from unittest.mock import Mock
 import pytest
 
 from ai_agent_toolbox.toolbox import Toolbox, ToolConflictError
-from ai_agent_toolbox.parser_event import ParserEvent, ToolUse
+from ai_agent_toolbox.parser_event import ParserEvent
+from ai_agent_toolbox.tool_use import ToolUse
 
 def test_use_tool_happy_path():
     """
